bonapp: Further cleanup
